combined PasswordService and UserService
